refactor: drop the _v2 C ABI tower and keep frequencies integral
The _v2 struct family existed only so hd_frequency.value could widen from
int32_t to double without breaking the existing C ABI. Nothing needs that
width: every frequency dictionary in circulation stores integers, so the
widening bought a parallel set of result structs and a duplicated
marshalling path for no reachable gain.

Frequency goes back to upstream's int value and non-optional
display_value, which removes hd_frequency_v2, hd_frequency_entry_v2,
hd_term_result_v2 and hd_lookup_result_v2, along with hd_query_run_v2 and
hd_lookup_run_v2. Neither of those two functions had a caller. The
templated dual marshalling in hoshidicts_c.cpp collapses to one concrete
path, and the string-frequency parsing that mirrored Yomitan's
_numberRegex goes with it.

hd_lookup_run_v3 is renamed hd_lookup_run_with_options and returns
hd_lookup_result, since there is no longer a v1/v2 pair for the version
suffix to distinguish. hd_lookup_options_v3 becomes hd_lookup_options.

What remains is the part that earns its place: the caller can choose
which frequency dictionary orders the results and in which direction, and
that ordering is applied before the max_results cap.

The delta against upstream include/hoshidicts_c.h is now purely additive
- one enum, one options struct and one function.

Also links glaze and unordered_dense into the test target, which could
not compile without them.

+++ src/lookup.cpp
@@ -21,18 +21,17 @@ std::vector<std::string> split_whitespace(const std::string& str) {
   return result;
 }
 
-std::optional<double> get_freq_value_for_dict(const TermResult& term, std::string_view dictionary_name,
-                                              bool descending) {
-  std::optional<double> frequency;
+std::optional<int> get_freq_value_for_dict(const TermResult& term, std::string_view dictionary_name, bool descending) {
+  std::optional<int> frequency;
   for (const auto& frequency_entry : term.frequencies) {
     if (frequency_entry.dict_name != dictionary_name || frequency_entry.frequencies.empty()) {
       continue;
     }
 
     for (const auto& candidate : frequency_entry.frequencies) {
-      frequency = frequency.has_value() ? std::optional<double>(descending ? std::max(*frequency, candidate.value)
-                                                                           : std::min(*frequency, candidate.value))
-                                        : std::optional<double>(candidate.value);
+      frequency = frequency.has_value() ? std::optional<int>(descending ? std::max(*frequency, candidate.value)
+                                                                        : std::min(*frequency, candidate.value))
+                                        : std::optional<int>(candidate.value);
     }
   }
